[gui] update clustering method

+++ libs/tracking/tract_model.hpp
@@ -43,7 +43,7 @@ public:
         // selection
         void delete_tracts(const std::vector<unsigned int>& tracts_to_delete);
         void select_tracts(const std::vector<unsigned int>& tracts_to_select);
-        void delete_repeated(double d);
+        void delete_repeated(float d);
         void delete_by_length(float length);
         unsigned int find_nearest(const float* trk,unsigned int length,bool contain = false);
